Fix: 댓글 더보기 연동 수정
